refactor/user: service내 패키지 세분화작업
