+++ src/main/java/min/taskflow/user/service/queryService/InternalUserService.java
@@ -1,4 +1,4 @@
-package min.taskflow.user.service;
+package min.taskflow.user.service.queryService;
 
 import lombok.RequiredArgsConstructor;
 import min.taskflow.user.dto.response.UserResponse;
@@ -7,6 +7,7 @@ import min.taskflow.user.exception.UserErrorCode;
 import min.taskflow.user.exception.UserException;
 import min.taskflow.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 
 /*
@@ -14,6 +15,7 @@ import org.springframework.stereotype.Service;
  */
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class InternalUserService {
 
     private final UserRepository userRepository;
