+++ src/test/java/min/taskflow/user/UserProfileTest.java
@@ -7,7 +7,7 @@ import min.taskflow.user.exception.UserErrorCode;
 import min.taskflow.user.exception.UserException;
 import min.taskflow.user.mapper.UserMapper;
 import min.taskflow.user.repository.UserRepository;
-import min.taskflow.user.service.ExternalUserService;
+import min.taskflow.user.service.queryService.ExternalUserService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
